Rearrange test and files (pull string out to its own file, move algebras to collections, consolidate conversion tests)
