+++ README.md
@@ -78,6 +78,20 @@ http://ramdajs.com/docs/#unless + lodash shorthand and f.callOrReturn support
 `f -> x -> f(find(f, x))`
 A version of `find` that also applies the predicate function to the result. Useful in gets
 
+## Collection Algebras or composable/recursive data types
+
+### map
+`map :: (a -> b) -> [a] -> [b]`
+Maps a function over an iterable. Works by default for Arrays and Plain Objects.
+
+### deepMap
+`deepMap :: (a -> b) -> [a] -> [b]`
+Maps a function over a recursive iterable. Works by default for nested Arrays, nested Plain Objects and mixed
+nested Arrays and Plain Objects. Also works for any other iterable data type as long as
+two other values are sent: a mapping function, and a type checker (See the
+unit tests for deepMap).
+
+
 ## Lodash Conversions
 These are conversions of lodash fp methods.
 
@@ -212,7 +226,9 @@ Flipped cascade
 ### cascadeKey
 A `_.get` that takes an array of paths and returns the first path that matched
 
+
 ## String
+
 ### parens
 `'asdf' -> '(asdf)'` Wraps a string in parenthesis.
 ### trimStrings
@@ -244,9 +260,11 @@ Maps `_.trim` through all the strings of a given object or array.
 
 Example: `('<b>', '<b>', [[0,1]], 'hi') -> '<b>h</b>i'`
 
+
 ## Math
 ### greaterThanOne
 `number -> bool` Returns true if number is greater than one.
+
 
 ## Lang
 Language level utilities
@@ -265,19 +283,6 @@ Returns true if the input has a `length` property > 1, such as arrays, strings, 
 
 ### append
 A curried, flipped `add`
-
-## Algebras or composable/recursive data types
-
-### map
-`map :: (a -> b) -> [a] -> [b]`
-Maps a function over an iterable. Works by default for Arrays and Plain Objects.
-
-### deepMap
-`deepMap :: (a -> b) -> [a] -> [b]`
-Maps a function over a recursive iterable. Works by default for nested Arrays, nested Plain Objects and mixed
-nested Arrays and Plain Objects. Also works for any other iterable data type as long as
-two other values are sent: a mapping function, and a type checker (See the
-unit tests for deepMap).
 
 
 ## Lens
