+++ test/conversions.spec.js
@@ -11,30 +11,59 @@ describe('Converted Functions', () => {
     bornAt: 'Thebes'
   }
 
-  it('getIn', () => {
-    expect(f.getIn(hero, 'name')).to.eql(_.get('name', hero))
+  describe('Flips', () => {
+    it('getIn', () => {
+      expect(f.getIn(hero, 'name')).to.eql(_.get('name', hero))
+      const obj = { a: 1 }
+      expect(f.inversions.getIn(obj)('a')).to.equal(1)
+      expect(f.getIn(obj)('a')).to.equal(1)
+    })
+  
+    it('pickIn', () => {
+      expect(f.pickIn(hero, 'name')).to.eql(_.pick('name', hero))
+      expect(f.pickIn(hero, ['name', 'father'])).to.eql(_.pick(['name', 'father'], hero))
+    })
+  
+    it('includesIn', () => {
+      let expectEql = (obj, name) => expect(f.includesIn(obj, name)).to.eql(_.includes(name, obj))
+      expectEql(hero, 'name')
+      expectEql(hero, 'Heracles')
+      expectEql(hero, 'Zeus')
+    })
   })
 
-  it('pickIn', () => {
-    expect(f.pickIn(hero, 'name')).to.eql(_.pick('name', hero))
-    expect(f.pickIn(hero, ['name', 'father'])).to.eql(_.pick(['name', 'father'], hero))
-  })
-
-  it('includesIn', () => {
-    let expectEql = (obj, name) => expect(f.includesIn(obj, name)).to.eql(_.includes(name, obj))
-    expectEql(hero, 'name')
-    expectEql(hero, 'Heracles')
-    expectEql(hero, 'Zeus')
-  })
-
-  it('extendOn', () => {
-    let expectEql = (clone, obj) => expect(f.extendOn(clone, obj)).to.eql(_.extend(obj, clone))
-    expectEql(_.clone(hero), { name: 'Hercules' })
-    expectEql(_.clone(hero), { consort: 'Auge' })
-  })
-
-  it('defaultsOn', () => {
-    let clone = _.clone(hero)
-    expect(f.defaultsOn(clone, { consort: 'Auge' })).to.eql(_.defaults({ consort: 'Auge' }, clone))
+  describe('Mutables', () => {
+    it('extendOn', () => {
+      let expectEql = (clone, obj) => expect(f.extendOn(clone, obj)).to.eql(_.extend(obj, clone))
+      expectEql(_.clone(hero), { name: 'Hercules' })
+      expectEql(_.clone(hero), { consort: 'Auge' })
+      expect(f.extendOn({
+        a: 1
+      }, {
+        a: 2,
+        b: 3,
+        c: 4
+      })).to.deep.equal({
+        a: 2,
+        b: 3,
+        c: 4
+      })
+    })
+  
+    it('defaultsOn', () => {
+      let clone = _.clone(hero)
+      expect(f.defaultsOn(clone, { consort: 'Auge' })).to.eql(_.defaults({ consort: 'Auge' }, clone))
+      expect(f.defaultsOn({
+        a: 2,
+        b: 3,
+        c: 4
+      }, {
+        a: 1
+      })).to.deep.equal({
+        a: 1,
+        b: 3,
+        c: 4
+      })
+    })
   })
 })
